+++ src/main/java/com/ecec/rweber/time/tracker/ActivityManager.java
@@ -11,15 +11,49 @@ import org.apache.log4j.Logger;
 import com.ecec.rweber.time.tracker.sql.DatasourceDrivers;
 import com.ecec.rweber.time.tracker.sql.SQLDatasource;
 import com.ecec.rweber.time.tracker.sql.SQLiteDatasource;
+import com.ecec.rweber.time.tracker.sql.upgrade.DatabaseUpgrade;
 import com.ecec.rweber.time.tracker.util.DBFile;
 
 public class ActivityManager {
+	private final int DB_VERSION = 1;
 	private Logger m_log = null;
 	private DBFile m_dbFile = null;  //manages storage of the DB file location
 	
 	public ActivityManager(){
 		m_log = Logger.getLogger(this.getClass());
 		m_dbFile = new DBFile();
+		
+		this.updateDatabase();
+	}
+
+	private void updateDatabase() {
+		SQLDatasource database = this.loadDatabase();
+		
+		int dbVersion = 0; //assume version 0
+		if(database.recordExists("select setting_name, setting_value from settings where setting_name = ?", "database_version"))
+		{
+			//check the version
+			List<Map<String,String>> versionQuery = database.executeQuery("select setting_name, setting_value from settings where setting_name = ?", "database_version");
+
+			dbVersion = Integer.parseInt(versionQuery.get(0).get("setting_value"));
+		}
+		
+		DatabaseUpgrade upgrade = null;
+		for(int i = dbVersion; i < DB_VERSION; i ++)
+		{
+			try {
+				upgrade = (DatabaseUpgrade) Class.forName("com.ecec.rweber.time.tracker.sql.upgrade.Version" + (i + 1)).newInstance();
+				
+				upgrade.doUpgrade(database);
+			}
+			catch(Exception e)
+			{
+				//can't recover from this
+				m_log.error("Error updating to Database Version: " + i);
+				e.printStackTrace();
+				System.exit(2);
+			}
+		}
 	}
 	
 	private SQLDatasource loadDatabase(){
@@ -37,7 +71,7 @@ public class ActivityManager {
 		
 		return result;
 	}
-	
+
 	private List<Activity> loadActivities(){
 		m_log.debug("loading activities");
 		SQLDatasource database = this.loadDatabase();
@@ -186,6 +220,13 @@ public class ActivityManager {
 	}
 	
 	public boolean setDatabaseLocation(File f) {
-		return this.m_dbFile.saveDatabaseLocation(f);
+		boolean result = this.m_dbFile.saveDatabaseLocation(f);
+		
+		if(result)
+		{
+			this.updateDatabase();
+		}
+		
+		return result;
 	}
 }