+++ src/main/java/com/ecec/rweber/time/tracker/sql/upgrade/DatabaseUpgradeException.java
@@ -1,0 +1,6 @@
+package com.ecec.rweber.time.tracker.sql.upgrade;
+
+public class DatabaseUpgradeException extends Exception {
+	private static final long serialVersionUID = 1L;
+
+}