+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 1.8.9
+
+### Added
+
+- implemented a DB update system to add settings table and default settings to older version automatically
+- added a settings table to the database to keep track of various system settings
+
 ## 1.8.8
 
 ### Changed